feat(login): implemented Login component
Implemented the Login component with required
click, change event handlers and JSX element
tree to render the intended UI.

The form submit event handler is incomplete.

The styled components to render JSX elements
for show password text and checkbox are to be
implemented and corresponding change event
handler is to be set for the checkbox element.