fix(state-management): updated context and file name
Updated the context created with React.createContext()
to include list of navigation link data that would be
used in the Header component.

Changed the const variable name storing the context to
AppContext and also the file name to AppContext.js.

It has been done to reflect the move of context data
and the AppContext.Provider component to the App component
from Header component. Having the context higher up in the
component tree will make the context data accessible to
child components down the tree.