refactor(App): added context component
Updated App component to manage state of
selected navigation link id, current app
theme chosen by user through their setters
and the context with AppContext.Provider
component with 'value' prop being passed
with the state and the list of navigation
link data.

This is the way to manage context data at
the root of the component tree, triggered
with the App component render through a
state change in other parts of the app.

It also allows access to context in child
components without having to pass on data
through props.