feat(header): implemented Header component
Implemented the Header component to render
the navbar with brand name, nav links and
account button.

The AppContext data is being accessed to
get the state of currently selected nav
link, user chosen app theme (light/dark),
list of nav links data. Accordingly, it is
being passed on to different components for
conditional styling.

Further work is to be done in rendering user
acoount information, logout action flow and
updating styled components to use the passed
in state and conditionally style the respective
components.