refactor: implemented higher order component
Implemented generic MappedPage component as a
higher order component. It essentially is the
blue print of the routed page layout, with header,
footer and body element passed as a prop through
the 'element' prop of Route component ( from
react-router-dom npm package) instance.

The Route component instances are part of the
router config (inside Routes component) setup
in the root App component.