refactor: added new higher order component
Implemented higher order component: ProtectedRoute,
to check for the existence of JSON Web Token with
the name: 'did-it-shrink-jwt-token', stored in
a cookie.

The cookie is being managed through js-cookie npm
package.

If the cookie renders an 'undefined' value for the
JSON Web Token (JWT), app will be redirected to 'login'
route and if a valid token is rendered, the requested
url will be navigated to the mapped element/component
to fetch the intended UI.

This will be used to check user authentication through
the JWT token in the cookie for below routes as per
the app design at the time of this commit:
(i) Home
(ii) Contribute
(iii) Leverage
(iv) Resources
(v) Not Found