fix(routing): updated routing paths
Fixed the incorrect path set for Navigate
component instance (imported from
react-router-dom npm package) rendered in
ProtectedElement component, to be absolute
path: '/login' and not relative path: 'login',
in order to ensure intended redirection to
Login component.

This redirection is done when user tries to
access different parts of the app without
logging in or user session expires.

Another route path fix has been done for Route
component instance that handles invalid route
request, matched with wildcard character: '*'.
So, the path set in this case is : '*' and
the invalid request redirected to the path:
'not-found', which is mapped to render NotFound
component.