fix(login): updated asynchronous code for jwt generation
Fixed the code for the successful creation
of JWT with asynchronous call to appwrite
backend api.

Moved the lines of code that redirect to
home route: '/', up on successful creation
and storage of JWT in a browser cookie, inside
the asynchronous success callback, along with
the line that sets the login status to success.

Earlier version had these lines outside the
asynchronous code and the redirection couldn't
pass the test that checks the existence of cookie
with JWT, thereby preventing redirection to home
route and rendering the intended UI.