fix(login): auth state and ui
Fixed the form input state management
in Login component to reset it when
user signup is successful and login
form is rendered for user to authenticate
with signed up credentials.

Added 'disable' attribute and value to
the LoginFormButton styled component, to
disable the button when the auth request
(signup/login) is being processed. That is
to prevent repetitive user clicks.