fix(login): updated auth and jwt creation logic
Fixed the incorrect creation of JSON
Web Token (JWT), even when no user
email or password was provided, thereby
incorrectly authenticating user and
logging one in to navigate the app.

Updated how the Appwrite backend API
is being used in the app with wrapper
functions, to return promises and handling
the success, failure scenarios in the Login
component. This includes the signup, login,
logout, JWT creation app flows.