feat(login): added form validation method
Implemented new form validation method
to process all form input fields, which
in turn calls validation method for
passed in field data.

This new method will be used in the Login
method and not the one that processes single
field data, as a way to process and display
error messages of different field inputs when
form is submitted, opposed to displaying the
error message on input change with the change
event handler. That is the design choice at
the time of this commit.