fix(utility): skip user's name validation for login
Updated code in validateFormInput() utility method
to skip the validation on user's name, with input
for login action, as only email and password would
be captured in the form.

Skipping user's name validation prevents false
positive case of failed validation, if name is
validated, which wouldn't have been the user
input for the login flow.