fix: input data for form validation
Updated the form data given as input
to the validateFormInput() utility
method, called in onFormSubmit() of
Login component.

The information on whether or not the
captured form input data is for login
or signup flow is also being passed in
now, to enable conditional validation
for user's name, which is absent for
login flow.