refactor: loader spinner utility
Created new utility folder: LoaderSpinner,
to include default configurations for any
loader spinners used in the app, imported
from react-loader-spinner npm package.

This commit includes code that sets the
default configuration for Oval loader spinner,
which will be exported in any component that
imports the LoaderSpinner utility. The named
import of the default config can then be used
in an instance of the react loader spinner
component, to set it up for rendering.