refactor(login): simplified loader spinner code
Imported default oval config from Loader Spinner
utility to be used in newly implemented method:
getStyledOvalSpinnerComponent(), in the Login
component, along with any custom config it gets
as input and return Oval loader spinner component
that is configured with custom styles.

The getStyledOvalSpinnerComponent() method has
code to use spread operator for creating the
final configuration as an aggregate of the default
and custom input configs, in a new js object.

That final config is in turn spread out as props
for the Oval loader spinner component, which is
returned by the method.

Also, the input custom config argument of the method
is set with a default empty object. So, to retain
the default configuration, no arguments would have to
be passed in with the call to the method:
getStyledOvalSpinnerComponent().