fix(auth): valid email regular expression
Fixed the incorrect regular expression set
with variable: validEmailRegExp, in
validateUserInput() method in Authentication
utility, by adding the preceding '\' to the
character class '\w', which matches [A-Za-z0-9_].

With the earlier version the installed Prettier
extension for VSCode editor was removing the '\'
from the entered '\w' in the constructor of RegExp,
with the 'new' keyword.

As a result, the intended match with the email input
from the form element of Login component was not
happening even with a valid email input.