feat(login): added loading spinner to submit button
Implemented code to add oval loading spinner
with custom configuration/styling, to
conditionally render the same when the auth
request is being processed.

The oval loading spinner is being fetched
through a call to the method:
getStyledOvalSpinnerComponent(), as a child
component instance inside the parent
LoginFormButton component, with text as Submit,
in the Login component JSX element tree.