styles: submit button loader spinner
Added new configuration property: wrapperStyle,
to the input configuration object for the call
to method getStyledOvalSpinnerComponent(), inside
the JSX code of LoginFormButton styled component
instance, within the Login component JSX element
tree.

The wrapperStyle property itself is an object, where
more customization can be done to the styles of the
rendered oval loader spinner, which is this case
is margin-right, to set some space between the
spinner and the button text.