feat(login): display auth error message
Implemented code to display the auth
request error message in Login
component.

The getSuccessMessageUI() method in
Login component has been renamed to
getAuthReqOutcomeMessageUI() and also
refactored to be a generic method to
display both success and failure
messages of auth request outcome, with
corresponding message configuration
as input, like the actual message, color
and react-icon component to be displayed
as part of the message in the Login
component JSX element tree.

The conditional code has been added to
display the error message set for: auth,
a property in errorMessages state variable.
The JSX element sub-tree of the message UI
will be added as a sibling, adjacent to the
LoginFormHeader styled component instance.
The conditional code has been added to
display the error message set for: auth,
a property in errorMessages state variable.
The JSX element sub-tree of the message UI
will be added as a sibling, adjacent to the
LoginFormHeader styled component instance.