styles(login): auth request outcome message
Adjusted CSS styling of the react-icon
and message displayed to indicate the
outcome of auth requests: signup, login.

Updated method getAuthReqOutcomeMessageUI()
to take additional property: iconSize, as
part of the input messageConfig, to have
different react-icon size for success and
failure messages to get the intended UI
in the Login component.