feat(header): implemented page header
Updated the Header component, rendered as
page header in the MappedPage higher order
component.

The code changes involve fixing numerous
styling aspects of the Header component
to render the intended UI, like:
(i) Styling of selected page that is
currently rendered, to have capsule like
border, white background with padding and
flipped text color relative to the background
color/text color combo of other page links
in the header menu.
(ii) Width of the header to fit the view-port
width, which involved removing the padding on
the left and right and avoiding the expansion
of the Header component width beyond the width
of the device view-port.
(iii) Height of the Header component and the
adjustment of font-sizes of text content in the
JSX elements.

Additionally, styling components have been added
to render list and list item of action buttons in
the Header component, like account and UI theme
(dark/light modes). Corresponding component
instances have been added to the Header component
JSX element tree.

Click event handlers are yet to be implemented and
set for the action buttons.