feat(header): changed react icon component
Updated the react icon component used
in the Header component to render
logout icon instead of the circle
account icon.

This is a temporary feature change
and will be reverted back to circle
account to display a menu with options
to update user account details, like,
user name, password, email and a
separate menu item for logging user out.