fix: jsx element box sizing
Added CSS selector * to the styles
of App component, to set the CSS
property value pair: 'box-sizing: border-box',
to include the padding in the JSX element
box size along with its content.

This is a way to avoid overflows and control
the sizing with height, width CSS properties.
With the default setting of box-sizing to
content-box, any padding added affects the
JSX element box size beyond the set width
and height and hence can overflow beyond the
device view-port width or height which will
accodringly create scroll bars to access the
content outside the view-port. As a result,
the rendered page layout and its content will
appear out of place an deviate from the
intended UI.