fix: header and footer sizing
Fixed the height and padding of styled components
in Header and Footer components to render the
intended UI, after the change in CSS box-sizing
property to border-box, done in App component
with the '*' CSS selector which applies to all
JSX elements in the App component tree.

With the earlier default setting of content-box,
the padding added to the outer components of Header
and Footer components resulted in content overflow
as the element box size was increased with padding,
on top of its content size (height, width). As a fix
for the overflow, left and right padding was removed
from the outer component elements and added to inner
child elements to render the intended UI with desired
spacing and layout.

That inner element padding has now been removed, which
had to be done at multiple places and padding could be
added to the outer element to get the desired spacing,
which also makes it intuitive and maintainable.