feat(footer): added jsx element tree
Added instances of styled components in
Footer component to create the JSX element
tree that renders the intended UI.

The AppContext has been used to access the
navLinkData and isLightTheme data, to render
nav links in the Footer component. Logic to
toggle styles between light and dark modes
is yet to be implemented.