feat(login): add header and footer to mapped login page
Updated the value of 'element' prop to the
child Route component instance in App component,
as part of the router config.

Wrapped the Login component with MappedPage
component, which is a higher order and generic
component with the Header and Footer components
rendered for every page mapped through the
react routing.