feat(mappedpage-hoc): add default empty method for nav link id setter
Added default empty method for the navLinkIdSetter
prop, in case nothing was passed in by the parent
during instantiation.

It was done to accommodate usage of MappedPage
higher order component for pages which are not
tracked or highlighted in the header as current
page and to still render the Header, Footer
components, along with the page content.

This fulfills the need for consistent branding
and styling across different pages rendered
as part of app control flows.