feat: add utility for nav link related app flow
Added new utility module with helper function
to determine user authentication state and
accordingly determine the nav link data to
be used to render JSX elements in App, Header
and Footer components.

This also helps with reuse of code/data for
new app features or extensions in the future,
enabling maintainability of increased complex
code base.