feat(login): set nav link data on successful authentication
Added code to store reference to nav link
data change callback and the call to the
same on successful login in onFormSubmit()
click event handler when login/signup form
is submitted.

The callback is actually stored in the
AppContext and accessed through the
AppContext.Consumer component instance
in the Login component. Once the callback
is invoked, the nav link data is set to
be used to render nav link JSX elements
in Header and Footer components.