feat(header): toggle brand name element with authentication
Added code to toggle or conditionally render
the brand name element in the Header component
based on user authentication state.

The brand name will be displayed for authenticated
page navigation flows, in the Header component, like
Contribute, Leverage, Resources. It will not be
displayed for Login, Join Now pages which are for
logging a user in and signing up a new user,
respectively.