refactor(authentication): create auth common component
Created new Auth common component to reuse
code for user login and new user registration
app flows.

It has been as part of redesigning app UI for
unauthenticated users to have separate Login
and Join-Now pages/routes for logging existing
users in and registering new users, instead of
the earlier design of single Login component
with the option to toggle between login and
signup forms depending on user needs.

The idea is to build Login, JoinNow components
on top of this common Auth component with
differentiating prop to trigger the appropriate
app flow.