fix(app): update getNavLinkIdSetter() method to work with nav link data
Updated the code in getNavLinkIdSetter() method
of App component which sets the selected nav link
id in the component state and also the AppContext
to be highlighted in the page Header component.

It was throwing an error with earlier implementation
which worked by directly accessing the id, without a
check on valid reference to nav link data object.
That threw an error of undefined reference to nav link
data for Route child component instances that were
currently not part of the active nav link data list,
set with data from the call to getPageNavLinkData()
method from the NavLinks utility module.

To fix it, the method getNavLinkIdSetter() has been
updated to take the nav link data object as input, to
do a check on undefined reference, rather that the id
object property itself, as its reference would fail if
the data object itself is undefined to begin with.