feat: add higher order component to style react icons
Added withStyling higher order component which
takes in a react icon component and returns a
new enhanced component with styling added to
the input component.

The imported IconContext from the react-icons
npm package has been used to style the child
instance of the input react icon component,
through the IconContext.Provider component
with CSS style properties passed as 'value'
prop.

This higher order component enables code
reusability to style multiple react icons at
different places in the app and to keep the
styling consistent across the app. It also
helps with better code maintainability with
changes done in one place being reflected in
component instances spread out across the app.

The component can be made more generic by adding
another input argument that captures the CSS styles.
That way, it can be used outside the BenefitsSection
component rendered on the home page, with the intended
styling. That is future work.