styles(benefitssection): adjusted css styling of styled components
Adjusted the CSS styling of properties in styled
components implemented to render JSX elements in
BenefitsSection child component instance of Home
component.

It has been as part of the iterative process of
evolving the benefits section UI to have intended
layout, border, content sizing, color palette
consistent with the primary color picked as part
of the app theme.