styles(home): adjust css styling in styled components
Adjusted the font size of HomeSectionHeader1 styled
component to be consistent with the header in the
BenefitsSection child component instance in the
Home component.

Added css height property to the HomeSectionContainer1
styled component with value conditionally set based on
input prop: sectionHeight, with default value of 60vh,
to have a uniform sizing for different section JSX
elements in the Home component.