refactor(home): import and add benefits section component
Replaced JSX element sub-tree rendered with
an instance of HomeSectionContainer2 styled
component and its children, with the
newly created and imported component
BenefitsSection, to render the content of
benefits section.

The new component has been created to enable
better maintainability and readability of the
Home component and also the growing code base.