feat: add data folder to organize app data
Created new data folder inside top level src
folder, to better organize data used to render
content in different parts of the app, which in
turn helps with managing the complexity of the
growing code base.

The benefitsData object which is used to store
the data about different benefits of using the app,
like icon, benefit name has been moved to the new
data folder inside a js file.

Also, the ImageData.js which was initially created
in src/assets/images folder has been moved to the
new data folder as well, to have all data in one
place.