feat: add contribute call to action section to be rendered in home component
Implemented new component ContributeCallToActionSection with
code to render the text and image content of the contribute
call to action section in the Home component. This section
serves the purpose of calling for action from the user
to contribute product data as part of the crowd sourcing
effort.

Corresponding image to be rendered as background of a container
element has been added to the src/assets/images folder. It has
been imported in src/data/ImageData.js file where imageData
object entry has been added for the same along with other
information like image attribution.

This data in turn has been imported in the new component and
passed on to the child component instance of
BackgroundImageWithAttribution to render the intended image
with attribution text.