refactor(home): import and instantiate new section components
Imported the newly refactored/created components that
render content of different sections and instantiated
child components, greatly simplifying the code in
Home component for better readability and maintainability.

The styled components also have been moved to the refactored
components or retained as common components to be imported
in different section components.