feat(background-image-with-attribution): add props to background image container styled component
Added expected props to the BackgroundImageContainer
styled component child instance in the BackgroundImageWithAttribution
component.

These props may or may not be passed as part of the
instantiation of BackgroundImageWithAttribution
component in the app component tree. Accordingly,
fail safe default values has been added to the
implementation of the BackgroundImageContainer
styled component.